Part 4: Regular Expressions
Last time we left off with a GUI for our keyword checker program. What
would be the next logical building block to add now that we have our
checker with a nice preview pane?

Well the preview pane is not too handy if it does not show the part that
we are looking for, so let us improve that today. We will use regular
expressions to achieve this. In short, a regular expression acts as a
text search, but instead of using a specific keyword to search for, it
is generally more effective with a pattern that finds a set of text
results that match this pattern. So our first piece of code is to add a
reference to System.Text.RegularExpressions at the top of our form code.

http://www.edvella.com/Post/23/Beginners_C_Part_4_Regular_Expressions